requirements.txt
mcp2210==0.1.4
numpy==1.26.3
pynput==1.7.6
pyserial==3.5
